Mqtt: fix handler not returning when initial value not received
